+++ zCore/build.rs
@@ -1,72 +1,8 @@
-use std::{env, fs::File, io::Write};
-
 fn main() {
-    if env::var("TARGET").unwrap().contains("riscv64") {
-        const TARGET_VADDR: usize = (!0) << 30;
-        // 不同的 BOOTLOADER 会将内核放在不同的位置
-        let kernel_base_addr = if env::var("PLATFORM").map_or(false, |p| p.contains("d1")) {
-            TARGET_VADDR + (1 << 20)
-        } else {
-            TARGET_VADDR
-        };
-
-        File::create("src/platform/riscv/linker.ld")
-            .unwrap()
-            .write_all(
-                format!(
-                    "\
-OUTPUT_ARCH(riscv)
-ENTRY(_start)
-BASE_ADDRESS = {kernel_base_addr:#x};
-{RISCV64_SECTIONS}"
-                )
-                .as_bytes(),
-            )
-            .unwrap();
-    } else if env::var("TARGET").unwrap().contains("aarch64") {
-        println!("cargo:rustc-env=USER_IMG=zCore/aarch64.img");
-    }
+    // 如果需要链接 rootfs 镜像，将镜像路径设置到环境变量
+    #[cfg(feature = "link-user-img")]
+    println!(
+        "cargo:rustc-env=USER_IMG=zCore/{}.img",
+        std::env::var("TARGET").unwrap()
+    );
 }
-
-const RISCV64_SECTIONS: &str = "
-SECTIONS
-{
-    . = BASE_ADDRESS;
-    start = .;
-
-    .text : {
-        stext = .;
-        *(.text.entry)
-        *(.text .text.*)
-        etext = .;
-    }
-
-    .rodata ALIGN(4K) : {
-        srodata = .;
-        *(.rodata .rodata.*)
-        *(.srodata .srodata.*)
-        erodata = .;
-    }
-
-    .data ALIGN(4K) : {
-        sdata = .;
-        *(.data .data.*)
-        *(.sdata .sdata.*)
-        edata = .;
-    }
-
-    .bss ALIGN(4K) : {
-        bootstack = .;
-        *(.bss.bootstack)
-        bootstacktop = .;
-
-        . = ALIGN(4K);
-        sbss = .;
-        *(.bss .bss.*)
-        *(.sbss .sbss.*)
-        ebss = .;
-    }
-
-    PROVIDE(end = .);
-}
-";